finish project calendar app

+++ src/hooks/useCalendarStore.js
@@ -1,30 +1,63 @@
 import { useDispatch, useSelector } from "react-redux"
-import { onAddNewEvent, onDeleteEvent, onSetActiveEvent, onUpdateEvent } from "../store";
+import Swal from "sweetalert2";
+import { calendarApi } from "../api";
+import { convertEventsToDateEvents } from "../helpers";
+import { onAddNewEvent, onDeleteEvent, onLoadEvents, onSetActiveEvent, onUpdateEvent } from "../store";
 
 
 export const useCalendarStore = () => {
     const dispatch = useDispatch();
     const { events, activeEvent } = useSelector(state => state.calendar);
+    const { user } = useSelector(state => state.auth);
 
     const setActiveEvent = (calendarEvent) => {
         dispatch(onSetActiveEvent(calendarEvent))
     }
 
     const startSavingEvent = async (calendarEvent) => {
-        //* TODO: llegar al BE
 
-        //* Happy path
-        if (calendarEvent._id) {
+        try {
             //* Actualizando
-            dispatch(onUpdateEvent({ ...calendarEvent }))
-        } else {
+            if (calendarEvent.id) {
+                await calendarApi.put(`/events/${calendarEvent.id}`, calendarEvent);
+                dispatch(onUpdateEvent({ ...calendarEvent, user }));
+                return;
+            }
+
             //* Creando
-            dispatch(onAddNewEvent({ ...calendarEvent, _id: new Date().getTime() }))
+            const { data } = await calendarApi.post('/events', calendarEvent);
+            dispatch(onAddNewEvent({ ...calendarEvent, id: data.evento.id, user }))
+        } catch (error) {
+            console.log(error);
+            Swal.fire('Error al guardar', error.response.data.msg, 'error')
+        }
+
+
+    }
+
+    const startDeletingEvent = async() => {
+        try {
+            if (activeEvent && activeEvent.id) {
+                await calendarApi.delete(`/events/${activeEvent.id}`);
+                dispatch(onDeleteEvent());
+                Swal.fire('Evento eliminado correctamente', '', 'success')
+            }
+        } catch (error) {
+            console.log(error)
+            Swal.fire('Error al eliminar evento', error.response.data.msg, 'error')
         }
     }
 
-    const startDeletingEvent = () => {
-        dispatch(onDeleteEvent())
+    const startLoadingEvents = async () => {
+        try {
+
+            const { data } = await calendarApi.get('/events')
+            const events = convertEventsToDateEvents(data.eventos);
+            dispatch(onLoadEvents(events));
+
+        } catch (error) {
+            console.log(error)
+        }
     }
 
     return {
@@ -37,6 +70,7 @@ export const useCalendarStore = () => {
         //* Metodos
         setActiveEvent,
         startSavingEvent,
-        startDeletingEvent
+        startDeletingEvent,
+        startLoadingEvents
     }
 }
